Remove PackageRow
