+++ src/app/components/resources/index.tsx
@@ -1,5 +1,3 @@
 export * from './ButtonRow';
-export * from './ExampleRow';
-export * from './PackageRow';
 export * from './ResourceRow';
 export * from './buttons';
